App/requirements.txt
flask
flask-login
flask-mail
flask-sqlalchemy
flask-security
python-dotenv
pytest
pytest-coverage
pytest-flask